Add test to ensure registrations persist
Simple test to ensure that new user registrations are persisted correctly in the database. Without this test a default Breeze app only tests to ensure a user is authenticated after logging in, and able to authenticate with already stored credentials. There is no test to ensure that a user's details are being persisted correctly during registration.