I am using PostgreSQL Database, and I have defined schemas. The validation that the user is unique fails!
I am using PostgreSQL Database, and I have defined schemas. For example, for user administration, the usuarios schema. Therefore the table to use to store the users would be  (usuarios.users).
So in my data model class, I have overwritten the getTable method so that it returns the name of the schema concatenated with the name of the table:
<?php

namespace App\Models;

use Illuminate\Contracts\Auth\MustVerifyEmail;
use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Foundation\Auth\User as Authenticatable;
use Illuminate\Notifications\Notifiable;
use Laravel\Sanctum\HasApiTokens;

class User extends Authenticatable
{
    use HasApiTokens, HasFactory, Notifiable;

    /**
     * The attributes that are mass assignable.
     *
     * @var string[]
     */
    protected $fillable = [
        'name',
        'email',
        'password',
    ];

    /**
     * The attributes that should be hidden for serialization.
     *
     * @var array
     */
    protected $hidden = [
        'password',
        'remember_token',
    ];

    /**
     * The attributes that should be cast.
     *
     * @var array
     */
    protected $casts = [
        'email_verified_at' => 'datetime',
    ];

    /**
     * Get the table associated with the model.
     *
     * @return string
     */
    public function getTable()
    {
        return env('DB_SCHEMA_USER') . '.users';
    }
}

With the changes that I have made to the controller, the validation works correctly, because the select count (*) from ..... is generated with the correct name of the schema + table.

Sorry but my English is not very good.